+++ tsrc/test/cli/test_status.py
@@ -16,7 +16,7 @@ def test_status_happy(tsrc_cli: CLI, git_server: GitServer, workspace_path:
     manifest_url = git_server.manifest_url
     tsrc_cli.run("init", manifest_url)
     tsrc.git.run(
-        workspace_path.joinpath("spam", "eggs"),
+        workspace_path / "spam/eggs",
         "checkout", "-b", "fish"
     )
 
@@ -32,7 +32,7 @@ def test_status_dirty(tsrc_cli: CLI, git_server: GitServer,
     git_server.push_file("foo/bar", "CMakeLists.txt")
     manifest_url = git_server.manifest_url
     tsrc_cli.run("init", manifest_url)
-    workspace_path.joinpath("foo", "bar", "CMakeLists.txt").write_text("DIRTY FILE")
+    (workspace_path / "foo/bar/CMakeLists.txt").write_text("DIRTY FILE")
 
     tsrc_cli.run("status")
 
@@ -47,7 +47,7 @@ def test_status_not_on_any_branch(tsrc_cli: CLI, git_server: GitServer,
     git_server.push_file("spam/eggs", "CMakeLists.txt")
     tsrc_cli.run("init", git_server.manifest_url)
     # corrupt the git
-    eggs_path = workspace_path.joinpath("spam/eggs")
+    eggs_path = workspace_path / "spam/eggs"
     tsrc.git.run(eggs_path, "checkout", "HEAD~1")
 
     tsrc_cli.run("status")
@@ -65,7 +65,7 @@ def test_status_on_tag(tsrc_cli: CLI, git_server: GitServer,
     manifest_url = git_server.manifest_url
     tsrc_cli.run("init", manifest_url)
     tsrc.git.run(
-        workspace_path.joinpath("spam", "eggs"),
+        workspace_path / "spam/eggs",
         "tag", "v1.0"
     )